requirements.txt
llama-cpp-python>=0.3.0
rocksdict>=0.3.20
faiss-cpu>=1.7.4
xxhash>=3.4.0
numpy>=1.24.0
prompt_toolkit>=3.0.40
cryptography>=41.0.0
psutil>=5.9.0
pyyaml>=6.0.0
rich>=13.0.0
